+++ GameLib/Food.cpp
@@ -6,7 +6,7 @@ Food::Food()
 {
 }
 
-Food::Food(int _id, sf::Vector2f _pos)
+Food::Food(int _id, sf::Vector2i _pos)
 {
 	id = _id;
 	position = _pos;
