+++ GameServer/ClientProxy.h
@@ -8,25 +8,25 @@
 class ClientProxy: public PlayerInfo
 {
 public:
-	ClientProxy::ClientProxy(int _id, std::string _alias, sf::IpAddress _ip, unsigned short _port, sf::Vector2f headPos);
+	ClientProxy::ClientProxy(int _id, std::string _alias, sf::IpAddress _ip, unsigned short _port, sf::Vector2i headPos);
 	~ClientProxy();
 
 	sf::IpAddress ip;
 	unsigned short port;
 	int numPings;
 
-	sf::Vector2f accumMovement;
+	sf::Vector2i accumMovement;
 
-	void CreateInitialBodyPositions(sf::Vector2f headPos);
+	void CreateInitialBodyPositions(sf::Vector2i headPos);
 	void PutBodyPositions(sf::Packet* pack);
 	void AddDataToPacket(sf::Packet* pack);
 
-	sf::Vector2f SumToHeadPosition();
+	sf::Vector2i SumToHeadPosition();
 	void CreateBodyPosition();
-	void UpdatePosition(sf::Vector2f headPos);
+	void UpdatePosition(sf::Vector2i headPos);
 	int lastIdMove;
 
 private:
-	const float separation = 10.0f;
+	const int separation = 10;
 };
